feat(Plug.Adapters.Cowboy2.Conn): make set-cookie header a list
In Cowboy 1.x, a list was used for the headers, having two set-cookie in
the headers would simply result in a duplicate header.

Now that Cowboy uses a map for the headers, the set-cookie must use a
list as there can only be one set-cookie header. Cowboy will map over
these when it converts the headers to a list.

https://tools.ietf.org/html/rfc7230#section-3.2.2
https://github.com/ninenines/cowboy/blob/ \
  3b91523a3c84a882dcc83f0591b919a939cb72c2/src/cowboy_http.erl#L903